// frontend/src/types.ts
/**
 * Тип книги.
 */
export interface Book {
	/**
	 * ID книги типа UUIDv4, используемый для идентификации книги в базе данных.
	 */
	id: string

	/**
	 * Название книги.
	 */
	title: string

	/**
	 * Год публикации книги.
	 */
	year: number

	/**
	 * Жанр книги.
	 *
	 * API возвращает названия жанров на английском языке, в UPPER CASE. Для конвертации жанра в русский язык воспользуйтесь `GENRE_NAMES_MAP`.
	 */
	genre: string

	/**
	 * Автор книги.
	 */
	author: string

	/**
	 * Количество просмотров книги.
	 */
	views: number
}